♻️ 마이 페이지 Info 컴포넌트 추상화
